V4/requirements.txt
# Housing V4 Requirements
torch>=2.0.0
torchvision
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.2.0
matplotlib>=3.5.0
seaborn>=0.11.0
PyYAML>=6.0
joblib>=1.2.0
